rename config/ccdlDatasetDownloadModal to config/ccdlDatasets, add learnMore object with learn more link attributes
