--- client/src/config/ccdlDatasetDownloadModal.js
@@ -1,14 +1,0 @@
-export const portalWideDatasets = {
-  ALL_METADATA: {
-    title: 'Portal-wide Sample Metadata'
-  },
-  SINGLE_CELL_SINGLE_CELL_EXPERIMENT: {
-    title: 'Portal-wide Data as SingleCellExperiment (R)'
-  },
-  SINGLE_CELL_ANN_DATA: {
-    title: 'Portal-wide Data as AnnData (Python)'
-  },
-  SPATIAL_SINGLE_CELL_EXPERIMENT: {
-    title: 'Portal-wide Spatial Data'
-  }
-}
